+++ src/formater.h
@@ -13,8 +13,8 @@ public:
     explicit Formater(Video const& video, CommandLine const& line);
     std::any GetFormat(const FORMAT eformat = FORMAT::Format) const;
 private:
-    int getClosest(int val1, int val2, int target) const;
-    int findClosest(int target) const;
+    int64_t getClosest(int64_t val1, int64_t val2, int64_t target) const;
+    int64_t findClosest(int64_t target) const;
 
     Video const& _video;
     CommandLine const& _line;
